testqwen/streamlit_qwen/requirements.txt
streamlit
torch
torchvision
torchaudio
torchao
git+https://github.com/huggingface/transformers.git
diffusers
Pillow
byaldi
qwen_vl_utils